requirements.txt
pydantic
requests
boto3
python-gnupg
pgpy
python-dateutil